requirements.txt
langchain-openai
langchain-core
langchain-text-splitters
langchain
requests
python-dotenv